Add a table encoding parameter for Netezza direct import
Direct import makes an external Netezza table using 'internal' encoding
for text colums. Then it integrates the external table into HDFS reading
it as a UTF-8 encoded stream.
But if the table contains VARCHAR that are not UTF-8 (e.g. ISO-8859-x),
the external table will share the same encoding as the source table, and
reading it as a UTF-8 encoded stream will corrupt non ASCII characters.